requirements.txt
aniso8601==9.0.1
click==8.1.3
dbus-python==1.2.16
distro-info==1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
python-apt==2.2.1
pytz==2022.6
six==1.16.0
unattended-upgrades==0.1
Werkzeug==2.2.2
zipp==3.11.0
